Clear cache and restart queue by default
Updating freescout doesn't differ from installing it, even if there is a
separate command `freescout:after-app-updat`[^1]. It's clear when you
take a look at the source code of this command[^2] that it does the same
things as needed during the initial setup. After pulling the new codea
migrations need to be run and the cache cleared. That does not differ
from those needed during setup. And since we would need to clear the
cache after every config update etc. At that point it would be required
to add a notify to every task.

In the future we could revisit the usage of hooks for performance
reasons, but first we should find mechanisms to detect an update.

Close #1

[^1]: https://github.com/freescout-helpdesk/freescout/wiki/Updating-FreeScout#method-3-update-using-git
[^2]: https://github.com/freescout-helpdesk/freescout/blob/master/app/Console/Commands/AfterAppUpdate.php#L38